+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.9.4
+- Add yes/no dialog for logout
 - Add settings for the password strength
 
 ## 0.9.3